refactor: minor refactors
