+++ src/deprecated files/presence.js
@@ -1,4 +1,4 @@
-const { botErrorHandler } = require('./botErrorHandler');
+const { botErrorHandler } = require('../utils/botErrorHandler');
 /**
  *
  * @deprecated
